func_jitterbuffer: fix audio failure caused by certain masquerade's.
During masquerade it is possible for the AST_JITTERBUFFER_FD to be
cleared (set to -1).  This caused RTP packets to be received late
and be discarded.  This change prevents masquerade from directly
modifying the jitterbuffer FD, and instead uses a fixup on the
datastore associated with the jitterbuffer to properly set the FD.

ASTERISK-22409 #close
Reported by: Corey Farrell

Change-Id: Ia0bfa02fc2d05bd11878cc4dd7c5a7f6c8acdb84